refactor(rsthreadedsim, rssim): Combine the simulation handling implementations
- The split of simulation handling across world, rsthreadedsim, and rssim was confusing and messy.
- New sim_threading implementation handles all simulation execution.
- Reduce inter-dependence between world, rsthreadedsim, and rssim.
- Add getters to World for protected members.

This commit improves the readability and logical flow of the simulation execution.
