+++ src/CMakeLists.txt
@@ -2,13 +2,13 @@ include_directories(${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_BINARY_DIR} ${CM
 
 set(fers_SRCS
 	logging.cpp main.cpp platform.cpp radar_system.cpp target.cpp world.cpp
-	xmlimport.cpp geometry_ops.cpp rspath.cpp rssim.cpp rsthreadedsim.cpp
+	xmlimport.cpp geometry_ops.cpp rspath.cpp sim_threading.cpp
 	rspulserender.cpp radar_signal.cpp rsparameters.cpp
 	antenna_factory.cpp response.cpp noise_generators.cpp rssignal.cpp
 	timing.cpp hdf5_export.cpp python_extension.cpp interpolation_set.cpp dsp_filters.cpp multipath_surface.cpp
 	antenna_pattern.cpp logging.h rspath.h portable_utils.h radar_system.h
-	rssim.h world.h geometry_ops.h platform.h target.h xmlimport.h
-	radar_signal.h rsparameters.h rssignal.h rsthreadedsim.h object.h
+	world.h geometry_ops.h platform.h target.h xmlimport.h
+	radar_signal.h rsparameters.h rssignal.h sim_threading.h object.h
 	antenna_factory.h response.h noise_generators.h timing.h hdf5_export.h python_extension.h interpolation_set.h
 	dsp_filters.h multipath_surface.h antenna_pattern.h
 	jones_vector.h
