--- src/rssim.h
@@ -1,14 +1,0 @@
-//rssim.h
-//Declarations and definitions for rssim.cpp and rsthreadedsim.cpp
-
-#ifndef RS_SIM_H
-#define RS_SIM_H
-
-#include "world.h"
-
-namespace rs
-{
-	void runThread(int threadLimit, World* world);  // TODO: unused function
-}
-
-#endif
