--- src/rsthreadedsim.h
@@ -1,11 +1,0 @@
-//rsthreadedsim.h
-//Definitions for threaded simulation code
-//Marc Brooker mbrooker@rrsg.ee.uct.ac.za
-//19 July 2006
-
-#ifndef RS_THREADEDSIM_H
-#define RS_THREADEDSIM_H
-
-#include "world.h"
-
-#endif
